"Unassigned" cannot be removed from selected filters
Resolves #897.
Resolves #879.

When filter value is assigned to NULL, this generates an error because the NULL case is not being handled.
If filter value is changed to something other than NULL, there is no safe way to distinguish between a case where "WHERE IS NULL" is needed and when not.

While it is highly unlikely that a person's name would be "Unassigned" or "None", I believe that it is better practice to ensure that this case is never an issue.
This is designed to utilize the NULL case as it appears to be intended to by furtherFilterByAssigneTo.html.
When handling the NULL case in this manner, the issue presented in #879 no longer becomes a problem.