Handle case where filterValue may be null when .toString() is used
This fixes the problem on the javascript/client-side because it is the least likely to introduce regressions.

On the java/server-side, org/tdl/vireo/controller/SubmissionListController.java in the function addFilterCriterion(), casting FilterValue to String is causing an error because java will not handle converting Boolean to a String.
This could be changed to use something like String.valueOf() when the filterValue is not null.