Provide error message and reset upload progress on file upload failures #855
There is already error messaging in place, but it is not implemented in every place.
Implement the errorMessage to be displayed in all of the file upload areas.

The error results from file upload failures is inconsistently structured.
Sometimes the errors are a string of HTML, other times the errors are objects.
This only parses the objects with known structures.

In all error cases, even if there is no parsed error message, an upload failure message is presented.