fieldValue.fileInfo is not always defined in the fieldValue for a file upload
When `fieldValue.fileInfo.uploaded` is checked, it is only checked to see if it is `undefined` so delete the property instead of setting the value to `false`.
The `fieldValue.fileInfo.uploaded` does not appear to be used but I have added it there for consistency.
I believe that the file upload (and respective failure functionality) should be moved to a utility or some other service/controller to be used more consistently.