Ludicrously named create submission repo normalization function breaks renaming labels (gloss)
The submission field profile repo create function is called only one time during runtime lifespan of the application.

It is worse in that this cache function seems to cache the names across all organizations.
The first organization with a created submission forces all organizations henceforth to use their submission properties.
For some reason, newly added fields are not affected by this.