Add support for HTML to be stored within the tooltip popup
I was unable to prevent `{{ }}` from filtering HTML, even when using the `trusted` filter that utilizes `$sceProvider`.
I was unable to use `ng-bind-html` because it needs to be applied to an attribute and not an entire html tag.

This leaves me with the undesirable literal addition of the html.

To reduce the potential problems, this is restricted to only definitions that add `"htmlTooltip":true,` to the appropriate Definition property.