Incorrect URL links when a custom context path at a subdirectory is used.
For example, if the context path is set to '/vireo', the home link of '/' would go to 'http://hostname/' instead of 'http://hostname/vireo'.
The HTML `<base>` attribute appears to be in use and set, so remove leading '/' in all cases to ensure the correct path is used.

This can be easily observed by:
1) Make sure the server is started with a custom context path such as '/vireo'.
2) Set a submission to 'Needs Correction' Status.
3) As a student, view the edit the submission and click the 'Corrections Completed' button and its respective confirmation.
4) The 'View submission status' has an incorrect link.

This can also be observed with clicking the "Home" link in the breadcrumb.