Implement repo test mocks (the skeleton structure)
This implements the skeleton structure of the repos as the first stage in implementing such a large volume of tests.
The base structure of the repo mocks is from the abstract repo as defined in weaver-ui-core.

Individual methods unique to an individual vireo repo are added but it appears there may be some overlap from the weaver ui core abstract repo and the vireo repo in some cases.
The mock repo data itself is not being populated in this pass unless it was already provided.

The repo data will be populated and the structure will be tested once model mocks and then controller tests are written.
Incomplete methods will be completed as needed or reviewed.