Implement model test mocks (the skeleton structure)
This implements the skeleton structure of the models as the first stage in implementing such a large volume of tests.
The base structure of the model mocks is from the abstract model as defined in weaver-ui-core.

Individual methods unique to an individual vireo models are added but may not be fully implemented or may simple be stub tests.
The mock model data itself is not being populated in this pass unless it was already provided.

The model data will be populated and the structure will be tested once the controller tests are written.
Incomplete methods will be completed as needed or reviewed.