Implement common/core angular mock tests
This simplifies the amount of code necessary to write mocks.
This has a caveat of requiring a slightly different way of implementing mocks.

The Repo mocks need to have the appropriate Model mocks (if applicable) included when the Repo mocks are included in a controller test.
For controller tests that expect Models to be passed should have the appropriate Model mock instead of the normal syntax (such as `mockUser' instead of `_User_`).
The Service mock for UserService requires the User model mock to be included for controller tests including UserService.

Tests that were disabled due to incompleteness have been enabled.
Some of tests exposed problems with the code.