Implement Tests for AbstractController, except for reportError() method
The reportError() method needs more work to get window.location.href tested without causing an actual redirect.
The use of a simulated context, such as:
```
var simulatedContext = {
    "window": {
        location: {
            href: ""
        }
    }
};

with (simulatedContext) {
  // ... do tests here ...
}
```

Prevents the redirect, but does not fully work when trying to test the results of $window.location.href.
More work is needed to get this specific test completed.