Refactor mock*1, mock*2, etc.. tests to data*1, data*2, etc..
The originating design used the mock*1 variables as model mocks.
With the current design, there is a mock* function that creates the files (and mocks the data via the mock() method).

Having both the function and the data objects begin with "mock" has increasingly become confusing.
Refactor the mock*1, etc.. models to instead be data*1 to improve readability.