Implement mockParameterModel() and fix usages of $q in controller tests
Passing an implementation of the mockModel abstract class to the controllers in place of the actual model is incorrect.
Instead, create a method (mockParameterModel) so that the correct model may be passed to the controller.
The correct model is one that handles an argument (that is not $q) for initialize the mock data.

For example,
```
  var first = new Degree();
  var second = new Degree(first);
```

Is the desired behavior a controller might perform.
If the mockDegree() is passed to the controller, `first` will be treated as `$q`.
This problem is avoided by returning a function that handles the data as a parameter.