Revert accidental inclusion of additional fix of adding missing method $scope.getSelectedOrganization
This issue should be in a separate commit.
The fix is for adding a missing method.
The reason for it to be missing is unclear in that it may be (somehow) being included dynamically outside of the controller.
A separate branch and PR will be used to address these kinds of issues.