Fix test issues with scope.getSelectedOrganization() being undefined
The template where OrganizationSettingsController is used ends up including OrganizationManagementController.
This results in additional methods on the scope before OrganizationManagementController is initialized.
OrganizationManagementController happens to be designed in such a way that the scopes provided by OrganizationSettingsController are required.

Change the tests to include the dependent controller such that the scope gets modified as desired prior to the initialization of OrganizationManagementController.