Improve test support for custom WsApi fetch mocking and improve test responses
The WsApi.fetch() call can now be mocked dynamically.
This allows for individual tests to cover different response values without having to hardcode anything in the WsApi mock implementation.

`valuePromise` in particular has been updated to support 'type' parameter such that a 'reject' or 'notify' may be used in addition to the default 'resolve'.