Universally implement initializeVariables() in tests
Move most of the variables, especially WsApi and $timeout into the newly created initializeVariables() function.
Most of the variables should not be re-defined when the initializeController() is called additional times.

This problem is most obvious in regards to $timeout.
To test functionality that happens only on controller initialization that requires $timeout a timeout has to be created before the controller initialization is called.
The previous behavior, the timeout would be reset during the controller initialization, breaking the link.
WsApi would, in theory, also be affected by this.