Embargo isActive property should be respected
Controlled Vocabularies reserve the ID property and are used in their database table.
The "identifier" property, however, is not distinct and can be safely utilized.
Embargo model should provide Embargo ID in "identifier" property, instead of the "duration" property when generating Controlled Vocabularies from Embargos.
(The duration property feels dangerous anyway, given that there may be more than one embargo with the same duration.)

The Controlled Vocabularies are processed in a very generalized way.
To make the Embargo respect this while also handling the isActive special case, a custom set of functions are passed along the scope.

The `showVocabularyWord()` method is used for checking to see if the Embargo (with the ID found in the "identifier" property of the Controlled Vocabulary) has isActive set to TRUE or FALSE.
In the cases where a Submission is available, make sure to always show a selected Embargo, even if it has isActive set to FALSE (for cases where isActive is set to FALSE after a submission already exists with the given Embargo).

The `displayVocabularyWord()` method is used to conditionally call the `showVocabularyWord()` method from within a given view or directive.
This will always return a boolean, and if `displayVocabularyWord()` is not provided, will always return TRUE.

Add a new directive called `DictionaryWords`.
The `fieldProfile` and `fieldValues` are not always available in the view in such a way that the filter can get both them and the "ng-repeat" value.
By providing a directive, the necessary properties can be placed on the scope and the ng-repeat can easily operate on the filter function.