Fix changing Embargo in Admin Submission Edit Page
The ng-click event is happening before the value changes, sending the wrong value to the database.
The ng-change event is now used.
Add additional logic to handle when the currently selected radio is clicked (no ng-change event is triggered).

Prevent save() from being executing while updating is in progress.
Add stopEditing() scope method to stop editing and close the edit mode when the currently selected value is clicked.
Add toggleFieldValue() to conditionally close the edit mode or save the field value based on what is clicked.