Issue 1525: Remove unique index on Submission, replacing it with a normal index.
Using `submissionDate` is not possible because it can be NULL and is NULL while the submission is In Progress.
A person intending to create two submissions for the same organization might not have at least one submitted.
Furthermore, the granularity of the `submissionDate` is per-day which is too large (it would need to include seconds at the least).

Replace the unique index with a normal index to at least maintain any performance gains of an index while dropping the uniqueness aspect.

The UI also needs changes to allow for creating a new submission.
The pre-existing behavior always tries to continue editing if an existing submission has the same organization for a new submission.
This makes it impossible to create a new submission for the same organization as an existing submission.
The if condition logic is changed such that only when the existing submissions organization no longer allows submissions shall it perform a "continue submission" rather than creating a new submission.

This introduces a new situation that did not exist before that needs additional consideration that is (partially) outside the scope of this change.
When there are two or more existing submissions for the same organization and that organization no longer allows submissions then there is way to select which of the existing submissions to continue.
This requires design change considerations.
Rather than make these decisions within the scope of this issue, I have opted to have this problem addressed in a separate issue.
The current implementation preserves the existing nuances where "continue" should still happen as it did before, so long as there is a single submission for some organization or that organization allows new submission.

The tests have been updating accordingly because the pre-existing tests are expecting uniqueness.
The `findBySubmitterAndOrganization()` method is only used by tests and has been update to `findAllBySubmitterAndOrganization()`.