Update: Improve angularjs tests for user.
The mock.user mock is not mocking in the correct way.
When a `new User()` command is called mocks fail.

Instead, have mock.user return a function, causing User to act as a constructor and allowing for `new User()`.
The original mockedUser must then be returned by the mocked constructor.

This exposed a bug in one of the tests, which was rewritten.

With this change, the `$injector.get()` no longer has problems and can be used.
It should be used and now is used for the repo tests.