Issue 1274: Redesign email handling to avoid race condition between Weaver and Spring.
There is a race condition during the loading data during sprint boot.

The `EmailSender` Bean may be loaded either before or after the Spring `@Value` annotations are loaded.
If `EmailSender` is loader after the `@Value` annotations are loaded, then the e-mails works as expected.
If the opposite happens, then no data is loaded and the Weaver default (which uses a mailinator address) is incorrectly used.

To avoid this, this commit does the following:
1) Do not use any `@Value` annotations (or any Spring configuration loaders) in the `AppEmailConfig` (where the Bean is loaded).
2) Implement a custom AppConfig loader, called `VireoEmailConfig`, to handle loading of the data.
3) Have the `SubmissionEmailService` Autowire the `VireoEmailSender` and use that to send the e-mails.
4) Load the settings for each e-mail (not ideal, future work is recommended to improve this).

The encoding format shouldn't be hard coded when everything else is not.
Furthermore, it seems that the Weaver design doesn't appear to set the encoding properly (Code shows `UTF-8` and actual e-mail headers show `us-ascii`).

The `VireoEmailConfig` now needs to be available for testing.
The Mockito `verify()` static method does a bad job of detecting if something actually is called.
Even though the final `sendEmail()` method that has 5 arguments gets called (thanks to adding the `doCallRealMethod()`), the `verify()` fails to recognize this.
The tests have to be changed to handle the specific function call even if it is just a pass through method.