1555: Email placeholders in Add Comment and Add File modals not working.
This is primarily a regression caused by the Lazy loading performance improvements.
Because the data is lazily loaded, it is not available and no e-mails are obtained to be sent.
Utilize a graph to force Eager loading in the explicit case of e-mails.
This required:
- Relocating controller logic code into the relevant service model and performing the database load there.
- Utilizing the attribute path entity graphs, see links below for details.
- have the controller instead pass the submission id onto the e-mail service methods.

Add the BCC to the e-mail debug log as well.

see: https://www.baeldung.com/spring-data-jpa-named-entity-graphs
see: https://www.baeldung.com/spring-open-session-in-view