1549: UI: Too many users, needs pagination to limit as well as a paginated table for "Who has Access" and "Add Member".
Change existing endpoints and add new endpoints.
- /assignable
- /assignable/total
- /unassignable
- /unassignable/total

These new endpoints now support pagination.

Rather than changing the UI internal design, I opted to just create a new endpoint for getting the totals.
These are needed for proper pagination support.

There is a work-around (see code comments) that allows for Pageable to have a default value of 0.
This means load all results without pagination.
In such a case, filtering is still supported.

In order to achive the filter, a synthetic DB column "name" is generated and is including at the DB level.
This is achieved by utilizing `@Formula` annotation.
This annotationr equired that I explicitly set the name annotation property for the firstName and lastName User properties.

Several JPA functions are added to accomplish all of the necessary tasks.

The UI loads the totals first and if the total is non-zero, then loads the actual data.

The existing filter from Add Member is removed in favor of the ng-table filter.

There is a lot more cleanup that could be done but for brevity and time-limits these potential improvements are ignored.

Problems that are not resolved:
- There is a flicker on filter changes or pagination that is annoying.
- Changes to either form still require full page refreshes.