Issue 1596: Add "expansive" option to CLI.
This "expansive" option, when enabled, should cause a large amount of action logs per submission (randomly).
This is intended to create more inconsistency in the generated data to better help identify performance problems due to large disparate data sets.

The action log now properly sets the submitter id on the log.

The generated submitter name now has a date and time stamp in the name to make it easier to more uniqely identify.