Issue 1596: Add action log debounce.
Set the debounce timeout to 2000 milliseconds, which is 2 seconds.
For the Reviewer Application page, set the timeout to 1000 milliseconds as this is expected to be directly used most often.

This utilizes the `::` (one time binding) AngularJs notation to ensure that the delay is processed exactly once.
There should be no reason for the delay to change with the current design.