Issue 1596: Fix major (DOS-level) regression regarding listeners on user repo.
This major (DOS-level) regression was introduced here in these commits:
- dc6973b61c39276d70949e2b0499d3009bfa8129
- adeb64122cc1e80c15c9c6eb0774c696c4265ec0
- aeb7bc97c08266fed98742da2ab99fc31b7330cd
- 4c346ef624237906c048f7019d379f96ae238ab6

Specifically, this Pull Request: https://github.com/TexasDigitalLibrary/Vireo/pull/1425

When any submission is created using the CLI, any open pages will freeze up and crash the browser when at least 2 submission are created.

Investigation shows numerous problems regarding listeners and how they are handled.
This specific fix focuses on problems regarding having listeners directly on the UserRepo.
Everything using the UserRepo will respond to any change on the UserRepo.
This then triggers re-loading the entire database (which is incredibly bad when the number of users is large).
Then subsequently triggers follow up POST and GET requests (2 POST and 2 GET request, for a total of 4).
Loading all users and posting data for every open browser for every tab will result in massive I/O spikes in both the service and the browser.
Usually the browser crashes first.
This happens on every page that has the UserRepo on it.

Any single operation on a user can cause the system or browsers crash.
Having two or more changes only exasperates the issue and makes the problem more obvious.

This change does the following:
- Strip out code that should not be on the repository model (specifically the listeners).
- Introduce the table-related functionality removed from the repository model onto the UserRepoController.
- Re-design small parts of the process to not rely on unnecessary information.
- Update the unit tests.
- Prevents this listener DOS on the browser and possibly the server and network.

This change does _not_ do the following:
- Add back the listeners logic on the controller.

This explicitly avoids adding the listeners back.
A new issue is recommended to provide a better implementation should listeners be desired on the user management page.