1596: Move custom action definition list read outside of the transaction.
Doing a select all read operation in a transaction write lock can have a performance hit.

When a student first loads a submission page and they do not have a submission, the submission gets created through this write method.
This causes a large pause on the view submission page.

Removing the select all read for the custom action definition list removes one potential cause of a slow first load.