Advisor review application page is always showing as rejected for generated submissions.
This is a follow up on the recent changes to the CLI for the purposes of generating more random data for performance testing.
The generated data is always setting the approvals to rejected.

This is happening because the date determines the approval boolean.
Change the design to only set approval of 30% of the submissions generated.
Of those with the approval set, have 50% be approved and 50% be rejected.

Fix relating typos and syntax problems.