Issue 1662: Improve accessibility regarding footer links, providing more details.
The links are dynamically generated and have three possibilities:
1) A link to an e-mail.
2) A link to a URL.
3) A self-link.

The message must be accurate to the purpose so programmatically generate the text.

In the third case, the description cannot be provided because there is no link anywhere.
I believe this may have been a design decision to preserve the link's presence.
In this specific case, the link has no purpose and the best way to communicate this is to tell the reader to ignore this element.

Ideally, one time binding would be used here as these values should not regularly change.
The one time binding does not appear to be supported for the `ng-attr` functionality.