Issue 1608: Add new end points for action log and re-design code depending on new end points.
Add two new end points:
1) `/submission/get-one/{submissionId}/action-logs`
2) `/submission/advisor-review/{advisorAccessHash}/action-logs`

These are placed within the submission controller as there is no action log controller and the path structure is within the submission path.
Two end points are created because the fetching by hash is not the only situation where the action logs are fetched for some submission.
The student submission view also has the same exact problem as with the advisor page.
The student submission view uses the submission ID as requires a logged in account whereas the advisor end point uses a hash and does not require a logged in user.

Two new matching repo methods are created to handle this behavior.
The by hash fetching is more advanced in that it needs to join submission table to properly match the hash.
I made several attempts to use the JPQL for `@Query` and I constantly ran into problems starting up.
I decided not to pursue this problem further for time-constraint reasons and move on.
I used native queries as a work-around, which is undesirable.

The UI action log table directive is updated to optionally use a callback rather than data straight off of the submission.
The pagination logic is passed directly through to the callback and the response is handled appropriately.
If the callback is not provided then the default behavior of processing the action logs off the submission is preserved.

Now that action logs are not always being loaded, in certain cases, additional undefined/NULL checks are required in the UI.

Unit tests are partially written.
There are no Submission Controller unit tests and so a completely new file is created and only the functions added here have tests written for them.
The repo methods are using `@Query` and so there is nothing to test for those methods.
The UI unit tests are written but are incomplete.
I see no existing pattern that has similar code fully working in a unit test and so a bare minimal test is written.

Implement a new JsonView called `SubmissionIndividualActionLogs`.
This allows for having a `SubmissionIndividual` without the action logs being lazily loaded.
The original behavior is preserved where needed by calling the newly minted `SubmissionIndividualActionLogs` instead of `SubmissionIndividual`.