Issue 1632: Allow file deletion to work even when the file on the disk is not found.
If the file no longer exists during the delete process then do not error out.
Instead, print a warning to the console.

There error actually first happens on `assetService.getAssetFileInfo()` which does not check to see if the file exists.
Modifying that code appears to have larger affects and is therefore not modified.
New functions are instead provided to allow for deletion without erroring on non-existent file.

Report that the file is not found in the action log where the file size would otherwise be.