Issue 1599, 1601: Dates, such as graduation dates, are stored incorrectly.
There are several problems causing this situation.
1. AngularJS Datepicker Problems
2. FieldProfileDirective Datepicker Problems
3. Work-around Datepicker Problems
4. System Input Types have Invalid Regex Patterns
5. Off-By-1-Day Problems
6. Incorrect Validation on Java Side

There are also some notes on migrating data:
1. Migrating Validators
2. Migrating Bad Data

### AngularJS Datepicker Problems
The AngularJS is not processing the datepickerOptions at all.
I've dived into the code and found it always uses full date and time with timezone and does conversion.
Given the state of AngularJS I expect there to be no fixes for their design decisions here.
The date formats are only being used for the label and not for the value.

### FieldProfileDirective Datepicker Problems
The Code setting the datepickerOptions is also flat out wrong, which further proves that these settings are being ignored.
For a Year+Month format, such as "MMMM yyyy", the maxDate cannot be month because it has a year!

### Work-around Datepicker Problems
Given the situation with the problems surrounding the Datepicker not doing what we want I can see only two obviois solutions.
1. Fork and fix or improve the AngularJS Datepicker and related modules (and also maintain them).
2. Implement a work-around of some sort.

I believe that 2 is the better choice at this time.

My particular approach in two is to just correct the data after we got it.
Because the value happens to be in a date time format, the value can be converted into what we want using the same exact filters via the Javascript date filter (via the AngularJS filter for this).

This introduces a problem in that Vireo uses a generalized design.
There is no obvious way to distinguish one predicate from another when it comes to what a valid value structure is.
However, the value of the predicate can be used to determine this.
As these are all configurable I decided that the simplest solution is to offload the logic into an appConfig property.
This reduces the need for code changes in the future should some new predicate need to be added.

I implemented to methods of matching the predicate:
1. start
2. exact

The first allows for matching a set of similar predicates, such as `dc.date.`.
The second allows for exact matching.
There is a potential third option, called 'regex', but I saw no reason to implement that at this time and I consider such a feature to be out of scope.

### System Input Types have Invalid Regex Patterns
The system input types in the SYSTEM_Input_Types.json are flat out wrong.