Issue 1599, 1601: More areas need date related changes.
The previous commit edcb52c6e750d82fa2aed26cfa4365e7f9d741c7 did not handle all of the cases it needed to.
(Also note that my message in the previous commit got truncated and is incomplete, see the relating PR for up to date details.)

The admin submission view and submission list view need to be updated.
I ended up needing to create a `dateColumns` in the appConfig as well.

Expand and alter the recently added appConfig settings to be more generalized (such as 'name' instead of 'predicate'), adding the 'format'.
There are still areas in the code where the format might be set, but I do not want to go too far with this change set.

With the data being stored in the expected format, the additional filters are not needed now.

The submission list java code needs to be updated to either no longer cast the data or to convert the data into a date rather than a date time.

Rewrite the related appConfig settings processing to find the value rather than to directly perform the action (of which the action is generally applying the format).

Preserve `INPUT_DATETIME` in case it is needed and so use `INPUT_DATE` instead to make the code less confusing and less innacurate.

This currently needs more work.
The Submission List dates are sometimes showing UNIX Timestamps or are not showing up at all.
This situation needs to be ivnestigated and fixed.

The CLI data generation needs to be updated to use "month" dates in certain cases.

The unit test is updated to remove a failing test for a function that no longer exists.
There are no unit tests added as of this commit but there needs to be.
More work will follow to implement the unit tests and updated existing affected unit tests.