Issue 1599, 1601: Change Database save structure and improve control over date handling.
Expand the build config settings adding custom database save format.
Use exact for 'dc.date.issued' rather than a start match of 'dc.date.' because it is not a normal date.

Update the CLI data generation to generate the data in the new format.
Improve the CLI date generation to be more accurate in terms of graduation semesters (degree date) and birth years.
That is, a semester has only 3 months that are valid and a birth year should have a year.

Update the Submission List SQL query.
- Add LOWER() to matches in regards to graduation semesters to avoid potential case-sensitivity problems in the month name.
- With the database data in proper SQL date formats, use the standard SQL CAST functionality.
- Move repeated code into a common private function.
- Remove INPUT_DATETIME as it is neither needed nor used.
- The filter strings are now in date format without time and so no special handling needs to be applied beyond casting to an SQL standard `DATE` type.

Looks like Graduation Month Validator is incorrectly changed.
In this case, it is a database table that stores a single numeric value and is not the same as the graduation semester (degree date).
Revert the change.

Update the template utility to work with the current dates.
- Detect if the type is INPUT_DATE and if so, then call the date formatter.
- Now that the date formatter is in use, the exception can be logged and do so as a warning.

The `INPUT_DATE` in the `SYSTEM_Input_Types.json` must be in database format for the validator but the error message must be in the displayed format.
This caused inconsistency in the message vs pattern.

The datepicker is messing up the date.
Work-around the datepicker doing this by stripping off the time and creating a new date object.
This effectively makes the object at UTC even though it is in local time.
Then pass this to the formatter where the time gets stripped off.
This gauarantees a consistent database date value.

There is still a problem with the datepicker on load.
Changes to address this shall follow this commit.

Be consistent with old behavior and hide the year view calendar in the case of graduation semester.
This reverts a change where I expanded the maxMode to `year` from the previous value of `month`.

A custom filter is used to handle displaying the field and parsing the display.
This is primarily used in the advisor review page.
Move related code into this filter and simplify the view page.
This filter is used in other places where the inputType is not necessarily available so also allow passing the type name as a string.